refactor: split up delete.rs
